Fixed favourite bug; fixed cart add bug

+++ app/src/main/java/com/example/final_project/data/remote/model/restaurant/RestaurantDetailsDto.kt
@@ -4,7 +4,7 @@ import com.squareup.moshi.Json
 
 data class RestaurantDetailsDto(
     val restaurantId: Int,
-    val deliveryFee: Double?,
+    val deliveryFee: Int?,
     val description: String,
     val image: String,
     val latitude: Double,
